// src/app/core/types/custom-values-top-image-card.ts
export type CustomValuesTopImageCard = {
    cardTopPart: {
        background: string,
        pokemonNameColorFont: string;
        labelsColorFont: string;
        badgesColorFont: string;
        colorBadges: string;
    },
    cardBottomPart: {
        background: string;
        statColorFont: string;
        radialProgressColor: string;
        radialProgressBackground: string;
        radialProgressColorFont: string;
        radialProgressCentralBackground: string;
    }
}